Learned about when to use *args and *kwargs
Learned about pd.crosstab() method in pandas to create pivot tables in python.

Explained the difference between *args and *kwargs to a 5 year old.

Learned to detect outliers using scatter plots and cross tables.

Learned how fuzzy matching algorithm can be in accurate because its not context and meaning aware. For example it can think a 'Preacher' is a 'Teacher' because the spelling closely match.

Learned to use list comprehensions to remove items from lists.